pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pinboard-mcp-server"
version = "0.1.0"
description = "Pinboard MCP Server - Read-only access to Pinboard bookmarks for LLMs"
authors = [
    {name = "Pinboard MCP Server Team"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastmcp>=2.0.0",
    "pinboard>=2.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "types-python-dateutil",
    "types-cachetools",
    "responses>=0.23.0",
    "vcrpy>=5.1.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/anthropics/pinboard-mcp-server"
Repository = "https://github.com/anthropics/pinboard-mcp-server"
Issues = "https://github.com/anthropics/pinboard-mcp-server/issues"

[project.scripts]
pinboard-mcp-server = "pinboard_mcp_server.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "B904", # check for raise without from inside except
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B018", "B019"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = false
disallow_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = false
warn_unused_ignores = false
warn_return_any = false
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pinboard.*",
    "vcrpy.*",
    "mcp.*",
    "fastmcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 90
